requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
httpx
beautifulsoup4
playwright
playwright-stealth
googlemaps
anthropic
google-auth
google-auth-oauthlib
google-api-python-client
apscheduler
python-dotenv
pydantic
pydantic-settings
dnspython
